[web] Detach component `CommentDrawer` from `Layout`
